Move presets into a config package
